# backend/docs/api.yaml
openapi: 3.0.3
info:
  title: RealTime Ticketing System API
  description: |
    API for managing real-time ticketing and case management with device integration, 
    WebSocket support, and Azure AD authentication.
    
    ## Authentication
    
    - **Session-based**: For staff/admin via Azure AD SSO
    - **Device-based**: For kiosk devices using device secrets
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /auth/login:
    get:
      tags:
        - Authentication
      summary: Initiate Azure AD login
      description: Redirects to Azure AD login page or returns 204 if already logged in
      responses:
        '302':
          description: Redirect to Azure AD login
        '204':
          description: Already logged in
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh session
      description: Validates and refreshes the current session
      responses:
        '200':
          description: Session refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '401':
          description: No valid session

  # Cases Routes
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sid
      description: Session-based authentication via cookies (for staff/admin users)
    
    deviceAuth:
      type: apiKey
      in: header
      name: Authorisation
      description: Device authentication using device secret (Bearer token)
...